Kernelloader 1.6 - Initrd is now supported even if kernel is not patched. - Initrd is now in protected memory area and can't get overwritten. - Clean up pad.c. - Code cleanup. - IOP heap bug fix. - Remove compiler warnings. - Add power off support. - Add mutex for some functions, to be sure that functions are called correctly. - Better initialisation of CDVD. - Fixed too short assembler loops. - Added SBIOS_DEBUG in config.mk for debugging SBIOS. - Removed redundant SIF DMA and SIF RPC structures. - Fixed lost instructions when debug is off (printf-if-combination).

+++ png2rgb/png2rgb.c
@@ -72,8 +72,8 @@ uint8_t *loadPng(const char *file_name, uint16_t * width, uint16_t * height,
 		PNG_TRANSFORM_SWAP_ENDIAN /* png_transforms */ , png_voidp_NULL);
 
 	/* At this point you have read the entire image */
-	printf("Width: %d Height: %d Depth: %d\n", info_ptr->width,
-		info_ptr->height, info_ptr->pixel_depth / 8);
+	printf("Width: %d Height: %d Depth: %d\n", (int) info_ptr->width,
+		(int) info_ptr->height, (int) (info_ptr->pixel_depth / 8));
 
 	*width = info_ptr->width;
 	*height = info_ptr->height;
